Dispatch HUD updates to the main thread
Fixes #30

SpriteKit seems to expect changes to happen on the main thread,
therefore we will schedule these changes there. Depending on the OS we
might already be on the main thread (e.g. on macOS), so we have to be
careful not to dispatch synchronously, since that would result in a
deadlock, therefore we submit the block asynchronously.